Add WordGameState sealed class
